orchestration_service/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
openai>=0.27.0
pydantic>=2.0.0
pydantic_settings>=2.0.3
requests>=2.28.0
python-dotenv>=0.21.0
pytest>=7.0.0
